add incident log section
