+++ src/logbook/components/SessionLogsTable.tsx
